+++ src/main/java/plugin/utils/Constants.java
@@ -1,0 +1,5 @@
+package plugin.utils;
+
+public class Constants {
+    public static String REPAINT_WHEEL_ON_FILE_CLOSING_EVENT = "repaintWheelOnFileClosing";
+}