+++ CHANGELOG.md
@@ -4,4 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## [1.0.0] - 2024-10-08
 ### Added
-- Used intellij-platform-plugin-template for gradle configuration and fixed compatibility issues.
+- Used intellij-platform-plugin-template for gradle configuration.
+
+### Fixed
+- Fixed compatibility issues.
+
+## [1.0.1] - 2024-10-14
+### Added
+- Refactored behavior so that the mouse cursor is no longer teleported to the center when closing a file.
+- Improved compatibility with other versions of the IDE.